+++ TestApplication/IOCPBuffer.h
@@ -8,15 +8,12 @@ namespace IOCP
 class IOCPBuffer
 {
 public:
-//	static const struct Flags
-//	{
-//	};
 
 	using BufferDelete = std::function<void(void *Buffer)>;
 
 	IOCPBuffer() noexcept;
-	IOCPBuffer(uint8_t *Buffer, size_t Size, OperationType Type, uint64_t SequenceNumber) noexcept;
-	IOCPBuffer(uint8_t *Buffer, size_t Size, OperationType Type, uint64_t SequenceNumber, BufferDelete Delete) noexcept;
+	IOCPBuffer(uint8_t *Buffer, uint32_t Size, OperationType Type, uint64_t SequenceNumber) noexcept;
+	IOCPBuffer(uint8_t *Buffer, uint32_t Size, OperationType Type, uint64_t SequenceNumber, BufferDelete Delete) noexcept;
 	IOCPBuffer(IOCPBuffer &) = delete;
 	IOCPBuffer(IOCPBuffer&& rhs) noexcept;
 	~IOCPBuffer() noexcept;
@@ -29,7 +26,7 @@ public:
 	bool Valid() const noexcept;
 	uint64_t SequenceNumber() const noexcept;
 	uint8_t *Pointer() const noexcept;
-	size_t Size() const noexcept;
+	uint32_t Size() const noexcept;
 	OperationType Type() const noexcept;
 	uint32_t Flags() const noexcept;
 	const IOCP_OVERLAPPED_EXTENSION *OverlappedExtension() const noexcept;
